# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOLVY Credit Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  timestamp: string;
  utilisateur: string;
  role: string;
  action: string;
  /** succes, echec */
  statut: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  ip_address?: string | null;
}